budget_pacing/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
